refactor gateway as a class
